Redesign of site
Total redesign of the site to add sections and multiple columns and clean
up the typography. Colours are still as per the original design for
consistency and the awesome 'Pair with me' badge features prominently.

A new 'Fork me on Github' banner has been added along with a new header
background, all other changes are limited to the layout/view files and
the scss stylesheet for the site.