Add Twitter code
Email isn't the only thing worth showing the code. With this you can
create a DM twitter intent.

I tried to keep it simple by having a simple button group that toggles
what will be produced when the submit button is pressed.

I was **unable** to test any of this due to
[lib v8](https://github.com/cowboyd/libv8/issues/123) problems so here
is the code and you can edit it as you wish.